fix: use dedicated gas limit for each tx in inspect_batch
