+++ crates/forge/tests/it/zk/factory_deps.rs
@@ -10,7 +10,6 @@ use foundry_test_utils::{
 use crate::{config::TestConfig, test_helpers::TEST_DATA_DEFAULT};
 
 #[tokio::test(flavor = "multi_thread")]
-#[ignore = "disabled since #476"]
 async fn test_zk_can_deploy_large_factory_deps() {
     let runner = TEST_DATA_DEFAULT.runner_zksync();
     {
@@ -19,20 +18,17 @@ async fn test_zk_can_deploy_large_factory_deps() {
     }
 }
 
-forgetest_async!(
-    #[ignore = "disabled since #476"]
-    script_zk_can_deploy_large_factory_deps,
-    |prj, cmd| {
-        util::initialize(prj.root());
+forgetest_async!(script_zk_can_deploy_large_factory_deps, |prj, cmd| {
+    util::initialize(prj.root());
 
-        prj.add_source(
-            "LargeContracts.sol",
-            include_str!("../../../../../testdata/zk/LargeContracts.sol"),
-        )
-        .unwrap();
-        prj.add_script(
-            "LargeContracts.s.sol",
-            r#"
+    prj.add_source(
+        "LargeContracts.sol",
+        include_str!("../../../../../testdata/zk/LargeContracts.sol"),
+    )
+    .unwrap();
+    prj.add_script(
+        "LargeContracts.s.sol",
+        r#"
 import "forge-std/Script.sol";
 import "../src/LargeContracts.sol";
 
@@ -43,39 +39,42 @@ contract ZkLargeFactoryDependenciesScript is Script {
     }
 }
 "#,
-        )
-        .unwrap();
+    )
+    .unwrap();
 
-        let node = ZkSyncNode::start();
+    let node = ZkSyncNode::start();
 
-        cmd.arg("script").args([
-            "--zk-startup",
-            "./script/LargeContracts.s.sol",
-            "--broadcast",
-            "--private-key",
-            "0x3d3cbc973389cb26f657686445bcc75662b415b656078503592ac8c1abb8810e",
-            "--chain",
-            "260",
-            "--gas-estimate-multiplier",
-            "310",
-            "--rpc-url",
-            node.url().as_str(),
-            "--slow",
-            "--evm-version",
-            "shanghai",
-        ]);
-        cmd.assert_success()
-            .get_output()
-            .stdout_lossy()
-            .contains("ONCHAIN EXECUTION COMPLETE & SUCCESSFUL");
+    // foundry default gas-limit is not enough to pay for factory deps in our current
+    // default environment
+    let gas_limit = u32::MAX >> 1;
 
-        let run_latest = foundry_common::fs::json_files(prj.root().join("broadcast").as_path())
-            .find(|file| file.ends_with("run-latest.json"))
-            .expect("No broadcast artifacts");
+    cmd.arg("script").args([
+        "--zk-startup",
+        "./script/LargeContracts.s.sol",
+        "--broadcast",
+        "--private-key",
+        "0x3d3cbc973389cb26f657686445bcc75662b415b656078503592ac8c1abb8810e",
+        "--chain",
+        "260",
+        "--gas-estimate-multiplier",
+        "310",
+        "--rpc-url",
+        node.url().as_str(),
+        "--slow",
+        "--gas-limit",
+        &gas_limit.to_string(),
+    ]);
+    cmd.assert_success()
+        .get_output()
+        .stdout_lossy()
+        .contains("ONCHAIN EXECUTION COMPLETE & SUCCESSFUL");
 
-        let content = foundry_common::fs::read_to_string(run_latest).unwrap();
+    let run_latest = foundry_common::fs::json_files(prj.root().join("broadcast").as_path())
+        .find(|file| file.ends_with("run-latest.json"))
+        .expect("No broadcast artifacts");
 
-        let json: serde_json::Value = serde_json::from_str(&content).unwrap();
-        assert_eq!(json["transactions"].as_array().expect("broadcastable txs").len(), 1);
-    }
-);
+    let content = foundry_common::fs::read_to_string(run_latest).unwrap();
+
+    let json: serde_json::Value = serde_json::from_str(&content).unwrap();
+    assert_eq!(json["transactions"].as_array().expect("broadcastable txs").len(), 3);
+});
